Fizalized

+++ business-logic-service/src/main/java/com/github/evseevda/stmlabstesttask/businesslogicservice/user/service/RoleService.java
@@ -5,7 +5,7 @@ import com.github.evseevda.stmlabstesttask.businesslogicservice.user.entity.Role
 
 public interface RoleService {
 
-    Role findRoleById(Long roleId);
+    Role findRoleById(Integer roleId);
     Role findRoleByName(String roleName);
 
 }
